Release 3.1.3

+++ SensorsAnalyticsSDK/Core/Builder/EventObject/SABaseEventObject.m
@@ -105,11 +105,11 @@
     id originalTime = self.properties[kSAEventCommonOptionalPropertyTime];
     if ([originalTime isKindOfClass:NSDate.class]) {
         NSDate *customTime = (NSDate *)originalTime;
-        NSInteger customTimeInt = [customTime timeIntervalSince1970] * 1000;
+        int64_t customTimeInt = [customTime timeIntervalSince1970] * 1000;
         if (customTimeInt >= kSAEventCommonOptionalPropertyTimeInt) {
             self.timeStamp = customTimeInt;
         } else {
-            SALogError(@"$time error %ld, Please check the value", (long)customTimeInt);
+            SALogError(@"$time error %lld, Please check the value", customTimeInt);
         }
     } else if (originalTime) {
         SALogError(@"$time '%@' invalid, Please check the value", originalTime);
