+++ SensorsAnalyticsSDK/Core/SAConstants.m
@@ -95,7 +95,7 @@ NSString * const kSAEventCommonOptionalPropertyProject = @"$project";
 NSString * const kSAEventCommonOptionalPropertyToken = @"$token";
 NSString * const kSAEventCommonOptionalPropertyTime = @"$time";
 //神策成立时间，2015-05-15 10:24:00.000，某些时间戳判断（毫秒）
-long long const kSAEventCommonOptionalPropertyTimeInt = 1431656640000;
+int64_t const kSAEventCommonOptionalPropertyTimeInt = 1431656640000;
 
 #pragma mark--lib method
 NSString * const kSALibMethodAuto = @"autoTrack";
